refactor: move assignments to it's own app
