+++ backend/ths/ths/settings.py
@@ -39,6 +39,7 @@ INSTALLED_APPS = [
     "rest_framework",
     "pets",
     "listings",
+    "assignments",
 ]
 
 MIDDLEWARE = [
